Scan A-Frame components. Scanners return warnings.
Adds first support for a non-polymer web components library: A-Frame! This patch will recognize certain kinds of invalid A-Frame component registrations. It also sets the groundwork for doing more interesting stuff later once we also recognize A-Frame entity definitions.

In order to make that work, I noticed that there was no way to return warnings from a scanner unless those warnings had a feature to attach to, so I refactored the return type of #scan().