Mark many fields readonly/immutable. Simplify element construction.
This is merely a descriptive change of how we already think of these fields.

More fields are still logically readonly/immutable, but we've got code that
initializes them in various places in a gradual way.

There is one functionality change in this PR, and that's that we treat behaviors
more like elements and more like mixins. We now mix behaviors into behaviors
and apply the same inheritance logic as mixins to behaviors.

This also makes element construction much more sane, by centralizing and documenting
how we deal with inheritance.