Fix reporting of internal error.
Likely related to https://github.com/Polymer/polymer-cli/issues/768

Previously we were casting an object to Warning, but this isn't valid. Warning is a class now and we depend on its methods.

Constructing a Warning in this internal error case was a little tricky, but this should work.

If this is what's behind https://github.com/Polymer/polymer-cli/issues/768 then we'll still need more work to track down the root cause, because this code path is one of those "this should never happen" situations, but at least it'll be clearer about what's happening and where.