Add UrlResolver#resolveFileUrl
This gives the Url Resolver control over how to go from a raw url found in a file to the final resolved url.

The default implementation does the same as we do now. This isn't hooked up anywhere yet. Hooking it up touches a ton of tests in a tedious way, so I split it up for reviewing.

First piece of https://github.com/Polymer/polymer-analyzer/issues/765