Assume Polymer properties are possibly null|undefined; add attributeType.
Polymer property definitions include a "type" field. This is a reference
to the constructor that will be used to deserialize attribute values.

In the absense of an explicit @type annotation, this deserializer is
also used by Analyzer to infer the Closure syntax property type.

Previously, we emitted one of: string|boolean|number|Object|Array|Date.
Note that in Closure syntax, some of these types are nullable, others
aren't (all non-primitive types in Closure are nullable by default), and
none can be undefined.

Now, we always emit a possibly null|undefined type when inferring the
property type.

We also now store the original attribute deserializer constructor name
in the "attributeType" field.